added max_client_string config variable to limit the length of the client version string in the login command
strlen() returns ssize_t so have to cast the Max_* to unsigned to avoid
compiler warnings

+++ topic.c
@@ -73,7 +73,7 @@ HANDLER (topic)
 	if (chan->topic)
 	    FREE (chan->topic);
 	/* if the topic is too long, truncate it */
-	if(Max_Topic > 0 && strlen(pkt) > Max_Topic)
+	if(Max_Topic > 0 && strlen(pkt) > (unsigned)Max_Topic)
 	    *(pkt+Max_Topic)=0;
 	if (!(chan->topic = STRDUP (pkt)))
 	{
